requirements.txt
uvicorn==0.21.1
fastapi==0.109.1
starlette==0.35.0
pydantic
databases==0.7.0
SQLAlchemy==1.4.47
aiomysql
httpx
redis
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
streamlit==1.40.0
numpy
pandas
matplotlib
seaborn
scikit-learn == 1.5.1
joblib
email-validator
python-dotenv
google-genai